requirements.txt
pandas>=2.0
xarray>=2024.1
netcdf4>=1.6
dask>=2024.1
erddapy>=2.1
copernicusmarine>=1.0
numpy>=1.26
tqdm>=4.66
pyyaml>=6.0
